Revised and systematised Meta
Overhaul of the structure and coding of the indicators dimension, in the light of the structure of the OECD data. New field introduced (Base Year, for constant, indexed and volume indicators). Not yet incorporated into OLTP and ROLAP table definitions because we wanted to push to Github at this point (after a lot of work)