OECD debugged
Offline dataset now contains total economy fixed assets from the OECD and Eurostat databases, which allows us to compare them. The OECD uses pre-SNA93 data, which is useful. Difference with Eurostat is for most cases around 2% but there are anomalies like the UK which may be caused by measure issues.

Have not yet checked the Eurostat and OECD capital consumption and net operational surplus estimates. (in the case of OECD it's not even clear these can be derived but it may be possible.