Retreival after gaffe with DimSource
Could not save workiing version because of internet availability. Started on next round of amendments and was unable to work back to a working version. Deleted all changes since last version and retrieved, hopefully.

This version has working OECD and Eurostat capital stock data which can be compared and tested.