US Capital Stock and Profit rate calculations
All calculations are in the 'DATA/OUTPUT' folder.
Source is in 'DATA/SOURCE/Masters/National Statistical Offices/USA' folder
Data not incorporated in the server build but calculations performed directly on  the NIPA and FED tables, which are very well organised.
Calculations are strictly provisional and offered on 'as is' basis; they need to be completed and then I will conduct diligence on them