US Definitive Source worksheet created
Created a definitive source workbook for the US based on national sources, for upload to the server. Before uploading, the aim is

(1) to validate and reconcile the standard series in this book

(2) to regularise the  extensive scratch calculations performed while interrogating the data

(3) to repeat at least for UK and possibly France and Spain

To this end a new folder 'VALIDATION' created where the previous scratch calculations are lodged for now. Aim is to standardise these.