Eliminated Pivoted intermediate tables
The project previously stored pivoted tables in intermediat waystation OLTP tables, which it then unpivoted into the fact file. This led to an unnecessarily cluttered layout.

In this commit, all input is directly to the fact file and each pivoted table is unpivoted as it is loaded.

The waystation files have not yet been eliminated, which is needed both to tidy up and to test that they really have been eliminated.

That will be done on the next commit