Tested IFS
See Change Tracking which also contains new list of proposed improvements.

Main change is that IFS has been tested, but some indicators in Balance of Payments are missing

Another change is that 'Hide Blank Fields' has been implemented so a lot of confusing and unnecessary additional indicators are no longer displayed. But this may have removed indicators from the display that are really needed, so this is Work in Progress