Greatly simplified dimensions
Tested somewhat perfunctorily, but with the old meta, ETL and SQL stashed away so they can be brought back into play if need be. From now on, the Geo and Indicator tables are held in 'standardised' form in the Meta.xlx spreadsheet. This does introduce a potential hazard in that the entries for any indicator that is supplied by multiple sources could potentially end up with different descriptions or metadata. However this would not break the system, merely make it potentially a bit confusing to the user. This trades off against a much simpler structure to the dataset.